move isValidUserEntry into top-most adapter class and have StandardComboboxAdapter implement choiceShouldBeShown() as well
