+++ app/src/org/commcare/adapters/StandardComboboxAdapter.java
@@ -2,8 +2,6 @@ package org.commcare.adapters;
 
 import android.content.Context;
 
-import org.commcare.views.widgets.SpinnerWidget;
-
 import java.util.Vector;
 
 public class StandardComboboxAdapter extends ComboboxAdapter {
@@ -20,29 +18,15 @@ public class StandardComboboxAdapter extends ComboboxAdapter {
     }
 
     @Override
-    public boolean isValidUserEntry(String enteredText) {
-        return isPrefixOfSomeChoiceValue(enteredText);
-    }
-
-    @Override
     public boolean shouldRestrictTyping() {
         return true;
     }
 
     /**
-     * This is the logic that the default Filter for an AutoCompleteTextView employs, so this
-     * is what the standard combobox adapter uses to determine if entered text is valid
+     * This is the logic that the default Filter for an AutoCompleteTextView employs
      */
-    private boolean isPrefixOfSomeChoiceValue(String text) {
-        for (String choice : choicesAllLowerCase) {
-            if (choice.startsWith(text.toLowerCase())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private boolean choiceShouldBeShown(String choice, CharSequence textEntered) {
-        return choice.toLowerCase().startsWith(textEntered.toString());
+    @Override
+    public boolean choiceShouldBeShown(String choice, CharSequence textEntered) {
+        return choice.toLowerCase().startsWith(textEntered.toString().toLowerCase());
     }
 }
